update snapshot_download, cache, delete no use python files, fix bug in train lora
